Remove unnecessary signature length check
This is implied by the requirement that the armored signature have at
least 64 bytes of base64 data.  If the first line is shorter than 64
bytes, either the overall armored signature will be too short, or the
check that a line shorter than 64 bytes is the last line before the CRC
will fail.