Check that the message has a trailing newline
If the command does not end with a newline, the subsequent split would
return an iterable of length 1, and its unpacking would raise a
ValueError.  This has no security impact, but it is still a bug.  Bad
input should not cause an uncaught exception.

Fix the problem by explicitly checking for a trailing newline, which
also provides a better error message than “Invalid end of signature
(trailing junk?)”.

Fixes: f2275c86325d ("Move signature rewrite into a separate function")