builder-github: update example config for commands
ALLOWED_DISTS_ is used in legacy builder in order to allow people
defined by their fingerprint to trigger github commands. This is
replaced for now by 'maintainers' field entry in github plugin
options. This is not the same 'maintainers' field that a component
would have. Indeed, in component, defining 'maintainers' field allow
to merge and use signed tag by a maintainer and not only by the
Qubes OS master key.