config: allow matching components by git URL
This will be useful for builder-github, which gets notification about
git repositories. With url_match=True, it can easily match repository to
a component (or components), even if the name doesn't match fully.

An example is `linux-kernel` repo, which is defined also as
`linux-kernel-latest` component, with the same repository (but different
branch).