Fix handling -rc versions
Do not strip -rc part of the version - it is necessary to download
correct tarball (`@VERSION@` placeholder in the URL). `-` in a version
will be invalid in some package formats (in rpm it is reserved as a
"release number" separator), so the packaging scripts need to handle it
properly in components where -rc versions are expected.

Furthermore, do not strip '-' from -rc suffix (as standard Version class
do). Do this by adding wrapper class (`QubesVersion`) that restores the
dash.