local: cleaup does need root access in some cases after all
If local executor is used for actual build with mock or pbuilder, it
will leave behind a bunch of root-owned files. Try to use sudo when
cleaning it up, and fallback to non-sudo call only if sudo one fails
(usually because the user doesn't have passwordless sudo access).

Partially reverts 39b72fe "local: no need to chown or remove as root
anymore"