Fix building the first arch package
Local repository on the first build is empty, but it still needs to have
metadata (qubes.db, qubes.files). `repo-add` used to create empty
metadata when called without any package, but since
https://gitlab.archlinux.org/pacman/pacman/-/commit/f91fa546f65af9ca7cdbe2b419c181df609969b7
it doesn't do it anymore. Since creating empty metadata is trivial
(based on the repo-add code), do that in the update-local-repo.sh script
directly.

This issue was especially hitting the CI, where every build is the first
package.