arch: workaround "GPGME error: Inappropriate ioctl for device" issue
GPGME used by pacman tries very hard to configure terminal if it thinks
it can. When running inside docker, it results in a series of errors
like "GPGME error: Inappropriate ioctl for device", which then are
interpreted by pacman as signature verification failure.

Workaround this by adding " | cat" to pacman calls (or rather - scripts
that call pacman) to make stdout a pipe instead of pts. If it's stupid
but it works, it ain't stupid, right?

Related to QubesOS/qubes-issues#9193
And to many failures in CI