installer: make lorax use only pre-verified repo
The "prep" stage fetches all the packages and verifies them. Make lorax
use only that repo, instead of feeding it also source online
repositories. This avoid the need to patch lorax to ensure it verifies
package signatures in all the cases and so we can use upstream lorax
directly.

QubesOS/qubes-issues#9402