Do not even schedule package jobs for components without packages
Do not schedule source/build/sign/publish jobs for components that have
packages disabled. This avoids unnecessarily evaluating their
dependencies. This is especially relevant for components that do have
dependencies, but are temporarily disabled by setting packages: false in
the config - for example when building dom0 packages, but without QWT.

And adjust tests to use a component with some packages for repository
create command.