Specify all job dependencies
Allow job dependencies with build=None for only ordering purpose, but
not implied effect on copy_in or check_dependencies. With that, it's
possible to specify ordering dependencies not only for build jobs, but
also for upload (that should execute after _all_ publish jobs).
Similar syntax can be later extended to resolve build=None into "all
build values of a component", but it's not necessary for now.

With that change, it's possible to specify dependencies for "upload"
jobs (that will depend on all packages), but also add missing
dependencies for "publish".