Do not schedule sign and publish jobs when not configured
Add is_signing_configured and is_publish_configured functions to the
distribution plugins, and use them to not schedule sign/publish jobs if
signing is not configured. This separation into two functions makes it
possible to have publish plugin that doesn't require signing - it simply
should provide own implementation of is_publish_configured() that
doesn't check for signing key configuration.

Simiarly, skip publish if target repository is not configured.
Technically, it's possible to use publish action without that config
(via 'repository publish' command), but it's a reasonable trade-off to
still require the setting, even if it gets overriden via cmdline later.
With this, we get to skip the job if not needed, avoiding an exception
in 'package all' command when signing is enabled but target repository
not.