Improve handling of fetch stage
skip-git-fetch is set to True by default.
Now fetch is enforced by specifically calling fetch command.
We trigger fetch stage in all the cases in order to ensure
that sources are available. This is mandatory to compute
all the dependencies notably to know if for a given distribution
there will be packages to be built later.